+++ YouTrackAutomation/YouTrackAutomation.psm1
@@ -15,10 +15,12 @@
 #Requires -Version 5.1
 Set-StrictMode -Version 'Latest'
 
-# Functions should use $moduleRoot as the relative root from which to find
+# Functions should use $moduleDirPath as the relative root from which to find
 # things. A published module has its function appended to this file, while a
 # module in development has its functions in the Functions directory.
-$moduleRoot = $PSScriptRoot
+$script:moduleDirPath = $PSScriptRoot
+
+$script:defaultIssueFieldDepth = 2
 
 # Attributes/properties for all [entitities](https://www.jetbrains.com/help/youtrack/devportal/api-entities.html) known
 # to be returned by functions in the module.Use by Get-YTEntityField to construct field query param values. If
@@ -300,7 +302,7 @@ $script:entityAttributes = @{
 # this file, so only dot-source files that exist on the file system. This allows
 # developers to work on a module without having to build it first. Grab all the
 # functions that are in their own files.
-$functionsPath = Join-Path -Path $moduleRoot -ChildPath 'Functions\*.ps1'
+$functionsPath = Join-Path -Path $script:moduleDirPath -ChildPath 'Functions\*.ps1'
 if( (Test-Path -Path $functionsPath) )
 {
     foreach( $functionPath in (Get-Item $functionsPath) )